mcp-servers/gemini-wrapper/requirements.txt
mcp[cli]>=1.0.0
google-generativeai>=0.8.0
pydantic>=2.0.0
